Rename to be explicit about getting by trust id
