+++ src/usecases/retrieveAverageVisitsPerDayByTrustId.js
